requirements.txt
Django==3.1.1
djangorestframework==3.11.1
djangorestframework_simplejwt==4.3.0
django-cors-headers==3.2.0
PyJWT==1.5.2
pytz==2017.2
requests==2.18.4
django-rest-swagger==2.2.0
urllib3==1.22
django-environ==0.4.5